@@ -1,7 +1,8 @@
 package tfar.speedrunnervshunter;
 
-import net.minecraft.item.Item;
-import net.minecraft.item.Items;
+
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.registries.ObjectHolder;
 
